mastermind/requirements.txt
requests
python-dotenv
PyGithub
dominate
feedparser
python-dateutil